Delete zombie file
